Refactor out ps command handler and add it to dotnet-dump
