+++ src/Tools/dotnet-trace/CommandLine/Commands/ListProcessesCommandHandler.cs
@@ -15,51 +15,14 @@ namespace Microsoft.Diagnostics.Tools.Trace
     {
         public static async Task<int> GetActivePorts(IConsole console)
         {
-            try
-            {
-                var processes = EventPipeClient.ListAvailablePorts()
-                    .Select(GetProcessById)
-                    .Where(process => process != null)
-                    .OrderBy(process => process.ProcessName)
-                    .ThenBy(process => process.Id);
-
-                foreach (var process in processes)
-                {
-                    try
-                    {
-                        Console.Out.WriteLine($"{process.Id, 10} {process.ProcessName, -10} {process.MainModule.FileName}");
-                    }
-                    catch (Exception)
-                    {
-                        Console.Out.WriteLine($"{process.Id, 10} {process.ProcessName, -10} [Elevated process - cannot determine path]");
-                    }
-                }
-
-                await Task.FromResult(0);
-                return 0;
-            }
-            catch (Exception ex)
-            {
-                Console.Error.WriteLine($"[ERROR] {ex.ToString()}");
-                return 1;
-            }
-        }
-
-        private static Process GetProcessById(int processId)
-        {
-            try
-            {
-                return Process.GetProcessById(processId);
-            }
-            catch (ArgumentException)
-            {
-                return null;
-            }
+            Console.Out.WriteLine(EventPipeClient.GetProcessStatus());
+            await Task.FromResult(0);
+            return 0;
         }
 
         public static Command ListProcessesCommand() =>
             new Command(
-                name: "list-processes",
+                name: "ps",
                 description: "Lists dotnet processes that can be attached to.",
                 handler: System.CommandLine.Invocation.CommandHandler.Create<IConsole>(GetActivePorts),
                 isHidden: false);
